secured API - request only with token

+++ api/general/delete.php
@@ -1,32 +1,51 @@
 <?php
-// Headers
+/**
+ * Created by PhpStorm.
+ * User: loukas
+ * Date: 11/27/2018
+ * Time: 5:10 PM
+ */
+
 header('Access-Control-Allow-Origin: *');
 header('Content-Type: application/json');
 header('Access-Control-Allow-Methods: DELETE');
 header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization,X-Requested-With');
 
+// database connection will be here
+// include database and object files
 include_once '../config/Database.php';
 include_once 'Solar.php';
-// Instantiate DB & connect
+include_once('User.php');
+
+// instantiate database and product object
 $database = new Database();
 $db = $database->connect();
 
-// Instantiate blog post object
-$solar = new Solar($db);
+$headers = apache_request_headers();
 
-// Get raw posted data
-$data = json_decode(file_get_contents("php://input"));
+if (isset($headers['Authorization'])) {
 
-// Set ID to UPDATE
-$solar->id = $data->id;
+    $user = new User($db);
 
-// Delete post
-if ($solar->delete()) {
-    echo json_encode(
-        array('message' => 'General deleted')
-    );
-} else {
-    echo json_encode(
-        array('message' => 'General not deleted')
-    );
+    if ($user->tokenExists($headers['Authorization'])) {
+
+        $solar = new Solar($db);
+
+        // Get raw posted data
+        $data = json_decode(file_get_contents("php://input"));
+
+        // Set ID to UPDATE
+        $solar->id = $data->id;
+
+        if ($solar->delete()) {
+            echo json_encode(
+                array('message' => 'General deleted')
+            );
+        }
+    } else {
+        header("HTTP/1.1 401 Forbidden");
+        $res = array("response" => "Login first");
+        echo json_encode($res);
+    }
 }
+
